+++ src/mandatory/philosopher/stop.c
@@ -6,7 +6,7 @@
 /*   By: tgrekov <tgrekov@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/27 07:06:01 by tgrekov           #+#    #+#             */
-/*   Updated: 2024/08/02 09:25:51 by tgrekov          ###   ########.fr       */
+/*   Updated: 2024/08/06 17:19:52 by tgrekov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,8 @@ int	stop(t_thread *thread)
 {
 	if (thread->err)
 		return (1);
-	if (timestamp() - thread->last_meal >= thread->global->opt.tt_die)
+	if (timestamp() - thread->last_meal
+		>= (unsigned long) thread->global->opt.tt_die)
 	{
 		if (!set_end(thread->global))
 			thread->err = status(thread, "died", 1);
